Adjusted variable Size

+++ src/TweetParsing.h
@@ -9,7 +9,7 @@ typedef struct _TWEET_PARSING_CONTEXT
 	FILE * File;
 	uint32_t FileID;
 	
-	uint8_t NumberOfDifferentUnicodes;
+	uint16_t NumberOfDifferentUnicodes;
 	UNICODE_APPEARANCE UnicodeAppearance[MAX_TWEET_CHARACTER_COUNT];
